Fixes
Moved XHTML Templates from SBMLPolisher to BiGGAnnotation
